feat(web): add latest cases to court page
